+++ src/src/command/new.rs
@@ -40,7 +40,7 @@ impl<'a> Command for NewCommand<'a> {
             match &self.current.target {
                 Some(target) => target.depth,
                 None => 0,
-            } as usize,
+            },
             self.current.target.as_ref().and_then(|t| t.parent_id),
             self.current.target.as_ref().and_then(|t| t.last_sibling_id),
         )]
